Refine sync logic; Init refactoring by adding a database service

+++ backend/cmd/dbPoc/main.go
@@ -1,18 +1,21 @@
 package main
 
 import (
+	"backend/pkg/constants"
+	"backend/pkg/enums/fileType"
 	"backend/pkg/models"
 	"backend/pkg/services/fileTreeManager"
+	"backend/pkg/services/imageConverter/utilities"
 	"context"
-	"database/sql"
+	"gorm.io/driver/postgres"
+	"gorm.io/gorm"
 	"log"
 	"os"
 	"path"
-
-	"gorm.io/driver/postgres"
-	"gorm.io/gorm"
+	"strings"
 )
 
+// TODO CONFIGURATION
 func main() {
 
 	connectionString := "host=localhost port=5432 user=streamingServiceAccount password=tempPassword dbname=streaming-db sslmode=disable TimeZone=Europe/Berlin"
@@ -20,14 +23,9 @@ func main() {
 	if openDatabaseConnectionError != nil {
 		log.Fatal(openDatabaseConnectionError)
 	}
-	// TODO TRANSACTION WHILE MIGRATING
-	db, dbError := databaseConnection.DB()
-	defer db.Close()
-	if dbError != nil {
-		log.Fatal(openDatabaseConnectionError)
-	}
 
-	addFileTypeEnumError := getAndExecuteSqlFile(db, "1_AddFileTypeEnum.sql")
+	ctx := context.Background()
+	addFileTypeEnumError := getAndExecuteSqlFile(databaseConnection, ctx, "1_AddFileTypeEnum.sql")
 	if addFileTypeEnumError != nil {
 		log.Fatal(addFileTypeEnumError)
 	}
@@ -39,9 +37,8 @@ func main() {
 
 	initializedFileTreeManager := fileTreeManager.NewFileTreeManager("E:\\projects\\videos-with-subtitle-player\\test-content").InitializeTree()
 	fileTreeFromDisk := initializedFileTreeManager.GetTree()
-	ctx := context.Background()
 
-	seedTagsError := getAndExecuteSqlFile(db, "2_tags_seed.sql")
+	seedTagsError := getAndExecuteSqlFile(databaseConnection, ctx, "2_tags_seed.sql")
 	if seedTagsError != nil {
 		log.Fatal(seedTagsError)
 	}
@@ -59,13 +56,16 @@ func main() {
 	log.Println(fileTreeItemsFromDb)
 }
 
-func getAndExecuteSqlFile(db *sql.DB, filename string) error {
+func getAndExecuteSqlFile(db *gorm.DB, ctx context.Context, filename string) error {
 	addFileTypeMigration, addFileTypeMigrationError := getMigrationFileContent(filename)
 	if addFileTypeMigrationError != nil {
 		return addFileTypeMigrationError
 	}
 
-	_, executeAddFileTypeMigrationError := db.Exec(addFileTypeMigration)
+	// TODO CREATE MIGRATION TABLE/EXECUTED SQL SCRIPTS
+	// TODO CHECKSUM FOR ALL EXECUTED SCRIPTS
+
+	executeAddFileTypeMigrationError := gorm.G[any](db).Exec(ctx, addFileTypeMigration)
 	if executeAddFileTypeMigrationError != nil {
 		return executeAddFileTypeMigrationError
 	}
@@ -73,6 +73,7 @@ func getAndExecuteSqlFile(db *sql.DB, filename string) error {
 	return nil
 }
 
+// TODO CHECK PATHS ON EXECUTION
 func getMigrationFileContent(filename string) (string, error) {
 	var relativeFilePathToMigrationFolder = path.Join("./migrations/", filename)
 	contentAsByteArray, readFileError := os.ReadFile(relativeFilePathToMigrationFolder)
@@ -122,6 +123,10 @@ func getDistinctFiles(left []models.FileTreeItem, right []models.FileTreeItem) [
 }
 
 func deleteFileTreeItemsFromDb(databaseConnection *gorm.DB, ctx context.Context, filesToDelete []models.FileTreeItem) error {
+	if len(filesToDelete) == 0 {
+		return nil
+	}
+
 	for _, fileToDelete := range filesToDelete {
 		_, deleteError := gorm.G[models.FileTreeItem](databaseConnection).Where("id = ?", fileToDelete.ID).Delete(ctx)
 		if deleteError != nil {
@@ -131,8 +136,9 @@ func deleteFileTreeItemsFromDb(databaseConnection *gorm.DB, ctx context.Context,
 
 		log.Printf("Successfully deleted file: %v\n", fileToDelete.Path)
 
-		if fileToDelete.AssociatedAudioFileId != nil {
-			doesAudioFileExits := doesFileWithFileIdExitsInDb(databaseConnection, *fileToDelete.AssociatedAudioFileId)
+		shouldTryCascadingDeletionOfAssociatedAudioFile := fileToDelete.AssociatedAudioFileId != nil
+		if shouldTryCascadingDeletionOfAssociatedAudioFile {
+			doesAudioFileExits := doesFileTreeItemWithFileIdExitsInDb(databaseConnection, *fileToDelete.AssociatedAudioFileId)
 			if doesAudioFileExits {
 				_, deleteError = gorm.G[models.FileTreeItem](databaseConnection).Where("file_id = ?", fileToDelete.AssociatedAudioFileId).Delete(ctx)
 				if deleteError != nil {
@@ -144,8 +150,9 @@ func deleteFileTreeItemsFromDb(databaseConnection *gorm.DB, ctx context.Context,
 			}
 		}
 
-		if fileToDelete.LowQualityImageId != nil {
-			doesLowQualityImageExist := doesFileWithFileIdExitsInDb(databaseConnection, *fileToDelete.LowQualityImageId)
+		shouldTryCascadingDeletionOfAssociatedLowQualityImage := fileToDelete.LowQualityImageId != nil
+		if shouldTryCascadingDeletionOfAssociatedLowQualityImage {
+			doesLowQualityImageExist := doesFileTreeItemWithFileIdExitsInDb(databaseConnection, *fileToDelete.LowQualityImageId)
 			if doesLowQualityImageExist {
 				_, deleteError = gorm.G[models.FileTreeItem](databaseConnection).Where("file_id = ?", fileToDelete.LowQualityImageId).Delete(ctx)
 				if deleteError != nil {
@@ -162,11 +169,76 @@ func deleteFileTreeItemsFromDb(databaseConnection *gorm.DB, ctx context.Context,
 	return nil
 }
 
-func insertFileTreeItemsIntoDb(databaseConnection *gorm.DB, ctx context.Context, filesToAdd []models.FileTreeItem) error {
-	// TODO this logic current can only create initial associations between items.
-	// TODO adding subtitle files (associatedAudioFiles) or images (lowQualityImage) later on (e. G. first creating a mp3 file and later on creating a vtt file)
-	// TODO will not be able to set corresponding associations (ids)
-	// TODO there has to be added
+func insertFileTreeItemsIntoDb(databaseConnection *gorm.DB, ctx context.Context, initialFilesToAdd []models.FileTreeItem) error {
+	if len(initialFilesToAdd) == 0 {
+		return nil
+	}
+
+	filesToAdd := make([]models.FileTreeItem, len(initialFilesToAdd))
+
+	/**
+	There are two cases files and their associated files can be added:
+		1. On the same batch: The associated file is added in the same batch from the file system as the file, so it is part of the input => do nothing
+		2. On a postposed batch: The associated file is not part of the same batch from the file system but could be already added to the database => try to get ids from the database
+
+	Only these types of files have this problem:
+		- subtitle
+		- image
+	*/
+	for i, initialFileToAdd := range initialFilesToAdd {
+		if initialFileToAdd.Type == fileType.IMAGE {
+			isLowQuality := utilities.IsLowQualityImage(initialFileToAdd.Path)
+			if isLowQuality {
+				pathWithoutLowQualitySuffix := utilities.RemoveLowQualitySuffixFromImageName(initialFileToAdd.Path)
+
+				isSameBatchInsert := false
+				for _, referenceFile := range initialFilesToAdd {
+					if referenceFile.Path == pathWithoutLowQualitySuffix {
+						isSameBatchInsert = true
+						break
+					}
+				}
+
+				if !isSameBatchInsert {
+					matchingImage, tryGetMatchingImageByPathError := tryGetFileTreeItemByPath(databaseConnection, ctx, pathWithoutLowQualitySuffix)
+					if tryGetMatchingImageByPathError != nil {
+						log.Printf("Warning: No matching audio file with path='%v' found. error='%v'\n", pathWithoutLowQualitySuffix, tryGetMatchingImageByPathError.Error())
+					}
+
+					if matchingImage != nil && matchingImage.ID > 0 {
+						initialFileToAdd.LowQualityImageId = &matchingImage.FileId
+					}
+				}
+
+			}
+		}
+
+		if initialFileToAdd.Type == fileType.SUBTITLE {
+			pathOfPossibleMatchingAudioFile := strings.TrimSuffix(initialFileToAdd.Path, constants.SubtitleExtension)
+
+			isSameBatchInsert := false
+			for _, referenceFile := range initialFilesToAdd {
+				if referenceFile.Path == pathOfPossibleMatchingAudioFile {
+					isSameBatchInsert = true
+					break
+				}
+			}
+
+			if !isSameBatchInsert {
+				matchingAudioFile, tryGetMatchingAudioFileByPathError := tryGetFileTreeItemByPath(databaseConnection, ctx, pathOfPossibleMatchingAudioFile)
+				if tryGetMatchingAudioFileByPathError != nil {
+					log.Printf("Warning: No matching audio file with path='%v' found. error='%v'\n", pathOfPossibleMatchingAudioFile, tryGetMatchingAudioFileByPathError.Error())
+				}
+
+				if matchingAudioFile != nil && matchingAudioFile.ID > 0 {
+					initialFileToAdd.AssociatedAudioFileId = &matchingAudioFile.FileId
+				}
+			}
+		}
+
+		filesToAdd[i] = initialFileToAdd
+	}
+
 	result := gorm.WithResult()
 	createInBatchError := gorm.G[models.FileTreeItem](databaseConnection, result).CreateInBatches(ctx, &filesToAdd, len(filesToAdd))
 	if createInBatchError != nil {
@@ -174,12 +246,16 @@ func insertFileTreeItemsIntoDb(databaseConnection *gorm.DB, ctx context.Context,
 		return createInBatchError
 	}
 
-	log.Printf("Created '%v' files\n", len(filesToAdd))
 	return nil
 }
 
-func doesFileWithFileIdExitsInDb(databaseConnection *gorm.DB, fileId string) bool {
+func doesFileTreeItemWithFileIdExitsInDb(databaseConnection *gorm.DB, fileId string) bool {
 	count := int64(0)
 	databaseConnection.Model(&models.FileTreeItem{}).Where("file_id = ?", fileId).Count(&count)
 	return count > 0
+}
+
+func tryGetFileTreeItemByPath(databaseConnection *gorm.DB, ctx context.Context, path string) (*models.FileTreeItem, error) {
+	matchingFile, err := gorm.G[*models.FileTreeItem](databaseConnection).Where("path = ?", path).First(ctx)
+	return matchingFile, err
 }
